Rename component: EmployerProfile -> CompanyProfile
